# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PromptGuard API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface AttackSimInput {
  /** One of: direct_injection, indirect_rag, multi_agent, jailbreak_roleplay, goal_hijacking, data_exfiltration */
  attackType: string;
  /** The attack payload text */
  payload: string;
}

export interface AttackSimResult {
  attackType: string;
  payload: string;
  /** What the vulnerable AI assistant responded */
  vulnerableResponse: string;
  attackSucceeded: boolean;
  /** What sensitive data or capabilities were exposed */
  dataExposed: string;
  analysisNotes: string;
}

export interface DefenseConfig {
  inputSanitization?: boolean;
  delimiterEnforcement?: boolean;
  intentClassifier?: boolean;
  leastPrivilege?: boolean;
  outputValidation?: boolean;
  auditLogger?: boolean;
}

export interface DefenseTestInput {
  attackType: string;
  payload: string;
  defenses: DefenseConfig;
}

export interface DefenseTestResult {
  payload: string;
  undefendedResponse: string;
  defendedResponse: string;
  /** 0-100 defense score */
  defenseScore: number;
  triggeredDefenses: string[];
  attackBlocked: boolean;
  /** @nullable */
  sanitizedPayload?: string | null;
}

export interface MessageAnalysisInput {
  message: string;
  conversationHistory?: string[];
}

export interface MessageAnalysisResult {
  injectionDetected: boolean;
  /** LOW, MEDIUM, HIGH, or CRITICAL */
  threatLevel: string;
  /** 0-100 confidence score */
  confidence: number;
  triggeredPatterns: string[];
  reasoning?: string;
}

export interface ChatMessage {
  role: string;
  content: string;
}

export interface RedTeamChatInput {
  message: string;
  /** vulnerable or hardened */
  mode: string;
  history?: ChatMessage[];
}

export interface RedTeamChatResult {
  response: string;
  mode: string;
}

export interface PlatformStats {
  attacksSimulated: number;
  defensesTested: number;
  vulnerabilitiesFound: number;
  activeDefenses: number;
}

export interface AuditEntry {
  id: string;
  timestamp: string;
  action: string;
  input: string;
  output: string;
  defenseTriggered: boolean;
  /** @nullable */
  details?: string | null;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface AnthropicConversationInput {
  title: string;
}

export interface AnthropicMessageInput {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}