# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PromptGuard API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /promptguard/simulate-attack:
    post:
      operationId: simulateAttack
      tags: [promptguard]
      summary: Simulate a prompt injection attack using Claude
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttackSimInput"
      responses:
        "200":
          description: Attack simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttackSimResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /promptguard/test-defense:
    post:
      operationId: testDefense
      tags: [promptguard]
      summary: Test defense mechanisms against a payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DefenseTestInput"
      responses:
        "200":
          description: Defense test result with side-by-side comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefenseTestResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /promptguard/analyze-message:
    post:
      operationId: analyzeMessage
      tags: [promptguard]
      summary: Analyze a chat message for injection attempts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageAnalysisInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageAnalysisResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /promptguard/chat:
    post:
      operationId: redTeamChat
      tags: [promptguard]
      summary: Send a message to the red-team agent (vulnerable or hardened mode)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedTeamChatInput"
      responses:
        "200":
          description: Agent response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedTeamChatResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AttackSimInput:
      type: object
      required: [attackType, payload]
      properties:
        attackType:
          type: string
          description: "One of: direct_injection, indirect_rag, multi_agent, jailbreak_roleplay, goal_hijacking, data_exfiltration"
        payload:
          type: string
          description: The attack payload text
    AttackSimResult:
      type: object
      required: [attackType, payload, vulnerableResponse, attackSucceeded, dataExposed, analysisNotes]
      properties:
        attackType:
          type: string
        payload:
          type: string
        vulnerableResponse:
          type: string
          description: What the vulnerable AI assistant responded
        attackSucceeded:
          type: boolean
        dataExposed:
          type: string
          description: What sensitive data or capabilities were exposed
        analysisNotes:
          type: string
    DefenseConfig:
      type: object
      properties:
        inputSanitization:
          type: boolean
        delimiterEnforcement:
          type: boolean
        intentClassifier:
          type: boolean
        leastPrivilege:
          type: boolean
        outputValidation:
          type: boolean
        auditLogger:
          type: boolean
    DefenseTestInput:
      type: object
      required: [attackType, payload, defenses]
      properties:
        attackType:
          type: string
        payload:
          type: string
        defenses:
          $ref: "#/components/schemas/DefenseConfig"
    DefenseTestResult:
      type: object
      required: [payload, undefendedResponse, defendedResponse, defenseScore, triggeredDefenses, attackBlocked]
      properties:
        payload:
          type: string
        undefendedResponse:
          type: string
        defendedResponse:
          type: string
        defenseScore:
          type: integer
          description: 0-100 defense score
        triggeredDefenses:
          type: array
          items:
            type: string
        attackBlocked:
          type: boolean
        sanitizedPayload:
          type: ["string", "null"]
    MessageAnalysisInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        conversationHistory:
          type: array
          items:
            type: string
    MessageAnalysisResult:
      type: object
      required: [injectionDetected, threatLevel, confidence, triggeredPatterns]
      properties:
        injectionDetected:
          type: boolean
        threatLevel:
          type: string
          description: "LOW, MEDIUM, HIGH, or CRITICAL"
        confidence:
          type: integer
          description: 0-100 confidence score
        triggeredPatterns:
          type: array
          items:
            type: string
        reasoning:
          type: string
    RedTeamChatInput:
      type: object
      required: [message, mode]
      properties:
        message:
          type: string
        mode:
          type: string
          description: "vulnerable or hardened"
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
        content:
          type: string
    RedTeamChatResult:
      type: object
      required: [response, mode]
      properties:
        response:
          type: string
        mode:
          type: string
...