# lib/api-zod/src/generated/types/attackSimInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PromptGuard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttackSimInput {
  /** One of: direct_injection, indirect_rag, multi_agent, jailbreak_roleplay, goal_hijacking, data_exfiltration */
  attackType: string;
  /** The attack payload text */
  payload: string;
}