# lib/api-zod/src/generated/types/attackSimResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PromptGuard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttackSimResult {
  attackType: string;
  payload: string;
  /** What the vulnerable AI assistant responded */
  vulnerableResponse: string;
  attackSucceeded: boolean;
  /** What sensitive data or capabilities were exposed */
  dataExposed: string;
  analysisNotes: string;
}