# lib/api-zod/src/generated/types/defenseTestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PromptGuard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DefenseConfig } from './defenseConfig';

export interface DefenseTestInput {
  attackType: string;
  payload: string;
  defenses: DefenseConfig;
}