# lib/api-zod/src/generated/types/defenseTestResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PromptGuard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DefenseTestResult {
  payload: string;
  undefendedResponse: string;
  defendedResponse: string;
  /** 0-100 defense score */
  defenseScore: number;
  triggeredDefenses: string[];
  attackBlocked: boolean;
  /** @nullable */
  sanitizedPayload?: string | null;
}