# lib/api-zod/src/generated/types/messageAnalysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PromptGuard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MessageAnalysisResult {
  injectionDetected: boolean;
  /** LOW, MEDIUM, HIGH, or CRITICAL */
  threatLevel: string;
  /** 0-100 confidence score */
  confidence: number;
  triggeredPatterns: string[];
  reasoning?: string;
}