# lib/api-zod/src/generated/types/redTeamChatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PromptGuard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface RedTeamChatInput {
  message: string;
  /** vulnerable or hardened */
  mode: string;
  history?: ChatMessage[];
}