# lib/api-zod/src/generated/types/redTeamChatResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PromptGuard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RedTeamChatResult {
  response: string;
  mode: string;
}